+++ src/modules/billing/controller.js
@@ -1,12 +1,10 @@
 'use strict';
 
 const Boom = require('@hapi/boom');
-const Event = require('../../lib/models/event');
 
 const { get } = require('lodash');
 const { envelope } = require('../../lib/response');
 const createBillRunJob = require('./jobs/create-bill-run');
-const prepareTransactionsJob = require('./jobs/prepare-transactions');
 const refreshTotalsJob = require('./jobs/refresh-totals');
 const { jobStatus } = require('./lib/event');
 const invoiceService = require('./services/invoice-service');
@@ -14,21 +12,10 @@ const invoiceLicenceService = require('./services/invoice-licences-service');
 const batchService = require('./services/batch-service');
 const transactionsService = require('./services/transactions-service');
 const billingVolumesService = require('./services/billing-volumes-service');
-const eventService = require('../../lib/services/events');
+const { createBatchEvent } = require('./lib/batch-event');
 
 const mapErrorResponse = require('../../lib/map-error-response');
 const mappers = require('./mappers');
-
-const createBatchEvent = (options) => {
-  const batchEvent = new Event();
-  batchEvent.type = options.type || 'billing-batch';
-  batchEvent.subtype = options.subtype || null;
-  batchEvent.issuer = options.issuer;
-  batchEvent.metadata = { batch: options.batch };
-  batchEvent.status = options.status;
-
-  return eventService.create(batchEvent);
-};
 
 /**
  * Resource that will create a new batch skeleton which will
@@ -198,35 +185,6 @@ const deleteInvoiceLicence = async (request, h) => {
   }
 };
 
-const postApproveReviewBatch = async (request, h) => {
-  const { batch } = request.pre;
-  const { internalCallingUser } = request.defra;
-
-  try {
-    const updatedBatch = await batchService.approveTptBatchReview(batch);
-
-    await billingVolumesService.approveVolumesForBatch(batch);
-
-    const batchEvent = await createBatchEvent({
-      type: 'billing-batch:approve-review',
-      status: jobStatus.processing,
-      issuer: internalCallingUser.email,
-      batch: updatedBatch
-    });
-
-    const message = prepareTransactionsJob.createMessage(batchEvent.id, updatedBatch);
-    await request.messageQueue.publish(message);
-
-    return envelope({
-      event: batchEvent,
-      batch: updatedBatch,
-      url: `/water/1.0/event/${batchEvent.id}`
-    });
-  } catch (err) {
-    return mapErrorResponse(err);
-  }
-};
-
 exports.getBatch = getBatch;
 exports.getBatches = getBatches;
 exports.getBatchInvoices = getBatchInvoices;
@@ -238,7 +196,6 @@ exports.deleteBatch = deleteBatch;
 
 exports.postApproveBatch = postApproveBatch;
 exports.postCreateBatch = postCreateBatch;
-exports.postApproveReviewBatch = postApproveReviewBatch;
 
 exports.patchTransactionBillingVolume = patchTransactionBillingVolume;
 exports.deleteInvoiceLicence = deleteInvoiceLicence;
