+++ src/modules/billing/services/volume-matching-service/matching-service.js
@@ -7,7 +7,6 @@ const { logger } = require('../../../../logger');
 const DateRange = require('../../../../lib/models/date-range');
 const ChargeElementGroup = require('./models/charge-element-group');
 const ReturnGroup = require('./models/return-group');
-const FinancialYear = require('../../../../lib/models/financial-year');
 
 /**
  * Allocates the return line volume
