+++ test/lib/connectors/repos/billing-invoice-licences.js
@@ -71,21 +71,6 @@ experiment('lib/connectors/repos/billing-invoice-licences', () => {
     });
   });
 
-  experiment('.findLicencesWithTransactionStatusesForBatch', () => {
-    let batchId;
-
-    beforeEach(async () => {
-      batchId = uuid();
-      await billingInvoiceLicences.findLicencesWithTransactionStatusesForBatch(batchId);
-    });
-
-    test('calls raw.multiRow with correct argumements', async () => {
-      const { args } = raw.multiRow.lastCall;
-      expect(args[0]).to.equal(queries.findLicencesWithTransactionStatusesForBatch);
-      expect(args[1]).to.equal({ batchId });
-    });
-  });
-
   experiment('.findOneInvoiceLicenceWithTransactions', () => {
     let result;
     beforeEach(async () => {
