+++ test/lib/connectors/repos/billing-volumes.js
@@ -65,43 +65,6 @@ experiment('lib/connectors/repos/billing-volumes', () => {
     });
   });
 
-  experiment('.findByChargeElementIdsAndFinancialYear', () => {
-    let result;
-    const ids = ['charge-element-id-1', 'charge-element-id-2'];
-
-    beforeEach(async () => {
-      result = await billingVolumes.findByChargeElementIdsAndFinancialYear(ids, 2019);
-    });
-
-    test('calls model.forge', () => {
-      expect(BillingVolume.forge.called).to.be.true();
-    });
-
-    test('queries for matching ID(s)', async () => {
-      const [operator, field, values] = stub.query.lastCall.args;
-      expect(operator).to.equal('whereIn');
-      expect(field).to.equal('charge_element_id');
-      expect(values).to.equal(['charge-element-id-1', 'charge-element-id-2']);
-    });
-
-    test('queries for matching financial year', async () => {
-      const [filter] = stub.where.lastCall.args;
-      expect(filter).to.equal({ financial_year: 2019 });
-    });
-
-    test('calls fetchAll', () => {
-      expect(stub.fetchAll.called).to.be.true();
-    });
-
-    test('calls toJSON() on returned models', async () => {
-      expect(model.toJSON.callCount).to.equal(1);
-    });
-
-    test('returns the result of the toJSON() call', async () => {
-      expect(result).to.equal({ foo: 'bar' });
-    });
-  });
-
   experiment('.update', () => {
     let result;
     const id = 'test-id';
