+++ test/modules/billing/routes.js
@@ -432,105 +432,6 @@ experiment('modules/billing/routes', () => {
     });
   });
 
-  experiment('getBatchLicences', () => {
-    let request;
-    let server;
-    let validBatchId;
-
-    beforeEach(async () => {
-      server = getServer(routes.getBatchLicences);
-      validBatchId = uuid();
-
-      request = {
-        method: 'GET',
-        url: `/water/1.0/billing/batches/${validBatchId}/licences`
-      };
-    });
-
-    test('returns the 200 for a valid payload', async () => {
-      const response = await server.inject(request);
-      expect(response.statusCode).to.equal(200);
-    });
-
-    test('returns a 400 if the batch id is not a uuid', async () => {
-      request.url = request.url.replace(validBatchId, '123');
-      const response = await server.inject(request);
-      expect(response.statusCode).to.equal(400);
-    });
-
-    test('contains a pre handler to load the batch', async () => {
-      const { pre } = routes.postApproveBatch.config;
-      expect(pre).to.have.length(1);
-      expect(pre[0].method).to.equal(preHandlers.loadBatch);
-      expect(pre[0].assign).to.equal('batch');
-    });
-  });
-
-  experiment('patchTransactionBillingVolume', () => {
-    let request, server, transactionId;
-
-    beforeEach(async () => {
-      server = getServer(routes.patchTransactionBillingVolume);
-      transactionId = '054517f2-be00-4505-a3cc-df65a89cd8e1';
-
-      request = {
-        method: 'PATCH',
-        url: `/water/1.0/billing/transactions/${transactionId}/volume`,
-        payload: {
-          volume: 5
-        },
-        headers: {
-          'defra-internal-user-id': 1234
-        }
-      };
-    });
-
-    test('returns the 200 for a valid payload', async () => {
-      const response = await server.inject(request);
-      expect(response.statusCode).to.equal(200);
-    });
-
-    test('returns a 200 if unknown headers are passed', async () => {
-      request.headers['x-custom-header'] = '123';
-      const response = await server.inject(request);
-      expect(response.statusCode).to.equal(200);
-    });
-
-    test('returns a 400 if the batch id is not a uuid', async () => {
-      request.url = request.url.replace(transactionId, '123');
-      const response = await server.inject(request);
-      expect(response.statusCode).to.equal(400);
-    });
-
-    test('returns a 400 if the calling user id is not supplied', async () => {
-      request.headers = {};
-      const response = await server.inject(request);
-      expect(response.statusCode).to.equal(400);
-    });
-
-    test('returns a 400 if the calling user id is not a number', async () => {
-      request.headers['defra-internal-user-id'] = 'a string';
-      const response = await server.inject(request);
-      expect(response.statusCode).to.equal(400);
-    });
-
-    test('returns a 200 if the volume is 0', async () => {
-      request.payload.volume = 0;
-      const response = await server.inject(request);
-      expect(response.statusCode).to.equal(200);
-    });
-
-    test('returns a 400 if the volume is not a positive number', async () => {
-      request.payload.volume = -5;
-      const response = await server.inject(request);
-      expect(response.statusCode).to.equal(400);
-    });
-
-    test('returns a 400 if the volume is omitted', async () => {
-      delete request.payload.volume;
-    });
-  });
-
   experiment('getInvoiceLicence', () => {
     let request;
     let server;
