+++ test/modules/billing/services/batch-service.js
@@ -20,7 +20,7 @@ const InvoiceLicence = require('../../../../src/lib/models/invoice-licence');
 const Licence = require('../../../../src/lib/models/licence');
 const Totals = require('../../../../src/lib/models/totals');
 const Transaction = require('../../../../src/lib/models/transaction');
-const { BatchStatusError, BillingVolumeStatusError } = require('../../../../src/modules/billing/lib/errors');
+const { BatchStatusError } = require('../../../../src/modules/billing/lib/errors');
 const { NotFoundError } = require('../../../../src/lib/errors');
 
 const eventService = require('../../../../src/lib/services/events');
@@ -96,8 +96,6 @@ experiment('modules/billing/services/batch-service', () => {
 
     sandbox.stub(transactionsService, 'saveTransactionToDB');
 
-    sandbox.stub(billingVolumesService, 'getVolumesWithTwoPartError').resolves([]);
-
     sandbox.stub(invoiceLicencesService, 'saveInvoiceLicenceToDB');
 
     sandbox.stub(invoiceService, 'saveInvoiceToDB');
@@ -118,6 +116,8 @@ experiment('modules/billing/services/batch-service', () => {
     sandbox.stub(newRepos.billingTransactions, 'deleteByInvoiceId');
     sandbox.stub(newRepos.billingInvoiceLicences, 'deleteByInvoiceId');
     sandbox.stub(newRepos.billingInvoices, 'delete');
+
+    sandbox.stub(billingVolumesService, 'approveVolumesForBatch');
   });
 
   afterEach(async () => {
@@ -971,13 +971,8 @@ experiment('modules/billing/services/batch-service', () => {
       });
 
       test('there are outstanding twoPartTariffErrors to resolve', async () => {
-        billingVolumesService.getVolumesWithTwoPartError.resolves([{ billingVolumeId: 'test-billing-volume-id' }]);
-        try {
-          await batchService.approveTptBatchReview(batch);
-        } catch (err) {
-          expect(err).to.be.an.instanceOf(BillingVolumeStatusError);
-          expect(err.message).to.equal('Cannot approve review. There are outstanding two part tariff errors to resolve');
-        }
+        billingVolumesService.approveVolumesForBatch.rejects();
+        expect(batchService.approveTptBatchReview(batch)).to.reject();
       });
     });
   });
