+++ test/modules/billing/services/charge-processor-service/index.js
@@ -12,7 +12,6 @@ const Transaction = require('../../../../../src/lib/models/transaction');
 const crmV2 = require('../../../../../src/lib/connectors/crm-v2');
 const chargeProcessorService = require('../../../../../src/modules/billing/services/charge-processor-service');
 const chargeVersionService = require('../../../../../src/modules/billing/services/charge-version-service');
-const transactionsProcessor = require('../../../../../src/modules/billing/services/charge-processor-service/transactions-processor');
 const batchService = require('../../../../../src/modules/billing/services/batch-service');
 const billingVolumeService = require('../../../../../src/modules/billing/services/billing-volumes-service');
 
@@ -128,7 +127,7 @@ experiment('modules/billing/services/charge-processor-service/index.js', async (
 
     sandbox.stub(chargeVersionService, 'getByChargeVersionId');
     sandbox.stub(batchService, 'getSentTPTBatchesForFinancialYearAndRegion');
-    sandbox.stub(billingVolumeService, 'getVolumes');
+    sandbox.stub(billingVolumeService, 'getVolumesForChargeElements').resolves([]);
   });
 
   afterEach(async () => {
@@ -149,6 +148,9 @@ experiment('modules/billing/services/charge-processor-service/index.js', async (
         chargeVersion = data.createChargeVersionWithTwoPartTariff();
         chargeVersionService.getByChargeVersionId.resolves(chargeVersion);
 
+        const billingVolumes = chargeVersion.chargeElements.map(data.createBillingVolume);
+        billingVolumeService.getVolumesForChargeElements.resolves(billingVolumes);
+
         // Run charge processor
         invoice = await chargeProcessorService.processChargeVersionYear(batch, financialYear, chargeVersion.id);
       });
@@ -165,9 +167,9 @@ experiment('modules/billing/services/charge-processor-service/index.js', async (
         )).to.be.true();
       });
 
-      test('the sent TPT batches for the financial year are loaded', async () => {
-        expect(batchService.getSentTPTBatchesForFinancialYearAndRegion.calledWith(
-          financialYear, batch.region
+      test('the billing volumes related to the charge elements in the charge version are fetched', async () => {
+        expect(billingVolumeService.getVolumesForChargeElements.calledWith(
+          chargeVersion.chargeElements, financialYear
         )).to.be.true();
       });
 
@@ -233,44 +235,6 @@ experiment('modules/billing/services/charge-processor-service/index.js', async (
         invoice.invoiceLicences[0].transactions.forEach(transaction => {
           expect(transaction instanceof Transaction).to.be.true();
         });
-      });
-    });
-
-    experiment('if transactions have isTwoPartTariffSupplementary flag set to true', () => {
-      let transaction, batch;
-      beforeEach(async () => {
-        chargeVersion = data.createChargeVersionWithTwoPartTariff();
-        chargeVersionService.getByChargeVersionId.resolves(chargeVersion);
-        transaction = data.createTransaction({ isTwoPartTariffSupplementary: true });
-        transaction.chargeElement.id = '00000000-0000-0000-0000-000000000000';
-        sandbox.stub(transactionsProcessor, 'createTransactions').returns([transaction]);
-        batch = data.createBatch('two_part_tariff');
-        invoice = await chargeProcessorService.processChargeVersionYear(batch, financialYear, chargeVersion.id);
-      });
-
-      test('the billingVolumes service is called with correct params', async () => {
-        const expectedChargeElement = [{
-          ...chargeVersion.chargeElements[1].toJSON(),
-          startDate: '2019-04-01',
-          endDate: '2020-03-31',
-          billableDays: 150,
-          authorisedDays: 150,
-          totalDays: 366
-        }];
-        const [chargeElement, licenceNumber, finYear, isSummer, batchArg] = billingVolumeService.getVolumes.lastCall.args;
-        expect(chargeElement).to.equal(expectedChargeElement);
-        expect(licenceNumber).to.equal(chargeVersion.licence.licenceNumber);
-        expect(finYear).to.equal(financialYear.yearEnding);
-        expect(isSummer).to.equal(true);
-        expect(batchArg).to.equal(batch);
-      });
-
-      test('does not include the charge element without a matching transaction', async () => {
-        const unrelatedElement = [
-          chargeVersion.chargeElements[0].toJSON()
-        ];
-        const [chargeElement] = billingVolumeService.getVolumes.lastCall.args;
-        expect(chargeElement).not.to.contain(unrelatedElement);
       });
     });
