+++ test/modules/billing/services/charge-processor-service/transactions-processor.js
@@ -7,14 +7,14 @@ const data = require('./data');
 experiment('modules/billing/services/charge-processor-service/transactions-processor', async () => {
   experiment('.createTransactions', async () => {
     const financialYear = data.createFinancialYear();
-    const sentTPTBatches = data.createSentTPTBatches();
     let chargeVersion, batch, transactions;
 
     experiment('for an annual batch', () => {
       beforeEach(async () => {
         batch = data.createBatch('annual');
         chargeVersion = data.createChargeVersionWithTwoPartTariff();
-        transactions = transactionsProcessor.createTransactions(batch, financialYear, chargeVersion, sentTPTBatches);
+        const billingVolumes = chargeVersion.chargeElements.map(data.createBillingVolume);
+        transactions = transactionsProcessor.createTransactions(batch, financialYear, chargeVersion, billingVolumes);
       });
 
       test('4 transactions are created', async () => {
@@ -58,7 +58,8 @@ experiment('modules/billing/services/charge-processor-service/transactions-proce
       beforeEach(async () => {
         batch = data.createBatch('supplementary');
         chargeVersion = data.createChargeVersionWithTwoPartTariff();
-        transactions = transactionsProcessor.createTransactions(batch, financialYear, chargeVersion, sentTPTBatches);
+        const billingVolumes = chargeVersion.chargeElements.map(data.createBillingVolume);
+        transactions = transactionsProcessor.createTransactions(batch, financialYear, chargeVersion, billingVolumes);
       });
 
       test('5 transactions are created', async () => {
@@ -110,7 +111,8 @@ experiment('modules/billing/services/charge-processor-service/transactions-proce
       beforeEach(async () => {
         batch = data.createBatch('two_part_tariff', { isSummer: true });
         chargeVersion = data.createChargeVersionWithTwoPartTariff();
-        transactions = transactionsProcessor.createTransactions(batch, financialYear, chargeVersion, sentTPTBatches);
+        const billingVolumes = chargeVersion.chargeElements.map(data.createBillingVolume);
+        transactions = transactionsProcessor.createTransactions(batch, financialYear, chargeVersion, billingVolumes);
       });
 
       test('1 transaction is created', async () => {
@@ -135,7 +137,8 @@ experiment('modules/billing/services/charge-processor-service/transactions-proce
           chargeVersion.chargeElements = [
             data.createChargeElement()
           ];
-          transactions = transactionsProcessor.createTransactions(batch, financialYear, chargeVersion, sentTPTBatches);
+          const billingVolumes = chargeVersion.chargeElements.map(data.createBillingVolume);
+          transactions = transactionsProcessor.createTransactions(batch, financialYear, chargeVersion, billingVolumes);
         });
 
         test('the transaction charge period is the full financial year', async () => {
@@ -164,7 +167,8 @@ experiment('modules/billing/services/charge-processor-service/transactions-proce
               }
             })
           ];
-          transactions = transactionsProcessor.createTransactions(batch, financialYear, chargeVersion, sentTPTBatches);
+          const billingVolumes = chargeVersion.chargeElements.map(data.createBillingVolume);
+          transactions = transactionsProcessor.createTransactions(batch, financialYear, chargeVersion, billingVolumes);
         });
 
         test('the transaction charge period is the full financial year', async () => {
@@ -194,7 +198,8 @@ experiment('modules/billing/services/charge-processor-service/transactions-proce
               }
             })
           ];
-          transactions = transactionsProcessor.createTransactions(batch, financialYear, chargeVersion, sentTPTBatches);
+          const billingVolumes = chargeVersion.chargeElements.map(data.createBillingVolume);
+          transactions = transactionsProcessor.createTransactions(batch, financialYear, chargeVersion, billingVolumes);
         });
 
         test('the transaction charge period is the full financial year', async () => {
@@ -217,7 +222,8 @@ experiment('modules/billing/services/charge-processor-service/transactions-proce
           chargeVersion.chargeElements = [
             data.createChargeElement()
           ];
-          transactions = transactionsProcessor.createTransactions(batch, financialYear, chargeVersion, sentTPTBatches);
+          const billingVolumes = chargeVersion.chargeElements.map(data.createBillingVolume);
+          transactions = transactionsProcessor.createTransactions(batch, financialYear, chargeVersion, billingVolumes);
         });
 
         test('the transaction charge period starts on the licence start date', async () => {
@@ -242,7 +248,8 @@ experiment('modules/billing/services/charge-processor-service/transactions-proce
           chargeVersion.chargeElements = [
             data.createChargeElement()
           ];
-          transactions = transactionsProcessor.createTransactions(batch, financialYear, chargeVersion, sentTPTBatches);
+          const billingVolumes = chargeVersion.chargeElements.map(data.createBillingVolume);
+          transactions = transactionsProcessor.createTransactions(batch, financialYear, chargeVersion, billingVolumes);
         });
 
         test('the transaction charge period ends on the earliest of expiry, lapsed and revoked dates', async () => {
@@ -263,7 +270,8 @@ experiment('modules/billing/services/charge-processor-service/transactions-proce
           chargeVersion.chargeElements = [
             data.createChargeElement()
           ];
-          transactions = transactionsProcessor.createTransactions(batch, financialYear, chargeVersion, sentTPTBatches);
+          const billingVolumes = chargeVersion.chargeElements.map(data.createBillingVolume);
+          transactions = transactionsProcessor.createTransactions(batch, financialYear, chargeVersion, billingVolumes);
         });
 
         test('the transaction charge period starts when the charge version starts', async () => {
@@ -284,7 +292,8 @@ experiment('modules/billing/services/charge-processor-service/transactions-proce
           chargeVersion.chargeElements = [
             data.createChargeElement()
           ];
-          transactions = transactionsProcessor.createTransactions(batch, financialYear, chargeVersion, sentTPTBatches);
+          const billingVolumes = chargeVersion.chargeElements.map(data.createBillingVolume);
+          transactions = transactionsProcessor.createTransactions(batch, financialYear, chargeVersion, billingVolumes);
         });
 
         test('the transaction charge period ends when the charge period ends', async () => {
@@ -305,7 +314,8 @@ experiment('modules/billing/services/charge-processor-service/transactions-proce
           chargeVersion.chargeElements = [
             data.createChargeElement({ timeLimitedStartDate: '2015-01-01', timeLimitedEndDate: '2016-01-01' })
           ];
-          transactions = transactionsProcessor.createTransactions(batch, financialYear, chargeVersion, sentTPTBatches);
+          const billingVolumes = chargeVersion.chargeElements.map(data.createBillingVolume);
+          transactions = transactionsProcessor.createTransactions(batch, financialYear, chargeVersion, billingVolumes);
         });
 
         test('no transactions are created', async () => {
@@ -320,7 +330,8 @@ experiment('modules/billing/services/charge-processor-service/transactions-proce
           chargeVersion.chargeElements = [
             data.createChargeElement({ timeLimitedStartDate: '2015-01-01', timeLimitedEndDate: '2025-01-01' })
           ];
-          transactions = transactionsProcessor.createTransactions(batch, financialYear, chargeVersion, sentTPTBatches);
+          const billingVolumes = chargeVersion.chargeElements.map(data.createBillingVolume);
+          transactions = transactionsProcessor.createTransactions(batch, financialYear, chargeVersion, billingVolumes);
         });
 
         test('the transaction charge period is the full financial year', async () => {
@@ -341,7 +352,8 @@ experiment('modules/billing/services/charge-processor-service/transactions-proce
           chargeVersion.chargeElements = [
             data.createChargeElement({ timeLimitedStartDate: '2019-05-01', timeLimitedEndDate: '2025-01-01' })
           ];
-          transactions = transactionsProcessor.createTransactions(batch, financialYear, chargeVersion, sentTPTBatches);
+          const billingVolumes = chargeVersion.chargeElements.map(data.createBillingVolume);
+          transactions = transactionsProcessor.createTransactions(batch, financialYear, chargeVersion, billingVolumes);
         });
 
         test('the transaction charge period starts on the time-limited start date', async () => {
@@ -362,7 +374,8 @@ experiment('modules/billing/services/charge-processor-service/transactions-proce
           chargeVersion.chargeElements = [
             data.createChargeElement({ timeLimitedStartDate: '2016-01-01', timeLimitedEndDate: '2019-06-01' })
           ];
-          transactions = transactionsProcessor.createTransactions(batch, financialYear, chargeVersion, sentTPTBatches);
+          const billingVolumes = chargeVersion.chargeElements.map(data.createBillingVolume);
+          transactions = transactionsProcessor.createTransactions(batch, financialYear, chargeVersion, billingVolumes);
         });
 
         test('the transaction charge period ends on the time-limited end date', async () => {
@@ -383,7 +396,8 @@ experiment('modules/billing/services/charge-processor-service/transactions-proce
           chargeVersion.chargeElements = [
             data.createChargeElement({ timeLimitedStartDate: '2019-05-01', timeLimitedEndDate: '2019-06-01' })
           ];
-          transactions = transactionsProcessor.createTransactions(batch, financialYear, chargeVersion, sentTPTBatches);
+          const billingVolumes = chargeVersion.chargeElements.map(data.createBillingVolume);
+          transactions = transactionsProcessor.createTransactions(batch, financialYear, chargeVersion, billingVolumes);
         });
 
         test('the transaction charge period starts and ends on the time-limited dates', async () => {
@@ -408,7 +422,8 @@ experiment('modules/billing/services/charge-processor-service/transactions-proce
           chargeVersion.chargeElements = [
             data.createChargeElement()
           ];
-          transactions = transactionsProcessor.createTransactions(batch, financialYear, chargeVersion, sentTPTBatches);
+          const billingVolumes = chargeVersion.chargeElements.map(data.createBillingVolume);
+          transactions = transactionsProcessor.createTransactions(batch, financialYear, chargeVersion, billingVolumes);
         });
 
         test('generate 2 charges', async () => {
@@ -431,7 +446,8 @@ experiment('modules/billing/services/charge-processor-service/transactions-proce
           chargeVersion.chargeElements = [
             data.createChargeElement()
           ];
-          transactions = transactionsProcessor.createTransactions(batch, financialYear, chargeVersion, sentTPTBatches);
+          const billingVolumes = chargeVersion.chargeElements.map(data.createBillingVolume);
+          transactions = transactionsProcessor.createTransactions(batch, financialYear, chargeVersion, billingVolumes);
         });
 
         test('generate 1 charge', async () => {
@@ -454,7 +470,8 @@ experiment('modules/billing/services/charge-processor-service/transactions-proce
           chargeVersion.chargeElements = [
             data.createChargeElement()
           ];
-          transactions = transactionsProcessor.createTransactions(batch, financialYear, chargeVersion, sentTPTBatches);
+          const billingVolumes = chargeVersion.chargeElements.map(data.createBillingVolume);
+          transactions = transactionsProcessor.createTransactions(batch, financialYear, chargeVersion, billingVolumes);
         });
 
         test('transactions have no agreements', async () => {
@@ -473,7 +490,8 @@ experiment('modules/billing/services/charge-processor-service/transactions-proce
           chargeVersion.chargeElements = [
             data.createChargeElement()
           ];
-          transactions = transactionsProcessor.createTransactions(batch, financialYear, chargeVersion, sentTPTBatches);
+          const billingVolumes = chargeVersion.chargeElements.map(data.createBillingVolume);
+          transactions = transactionsProcessor.createTransactions(batch, financialYear, chargeVersion, billingVolumes);
         });
 
         test('transactions have the agreements applied', async () => {
@@ -494,7 +512,8 @@ experiment('modules/billing/services/charge-processor-service/transactions-proce
           chargeVersion.chargeElements = [
             data.createChargeElement()
           ];
-          transactions = transactionsProcessor.createTransactions(batch, financialYear, chargeVersion, sentTPTBatches);
+          const billingVolumes = chargeVersion.chargeElements.map(data.createBillingVolume);
+          transactions = transactionsProcessor.createTransactions(batch, financialYear, chargeVersion, billingVolumes);
         });
 
         test('4 transactions are created', async () => {
