+++ test/modules/billing/services/volume-matching-service/matching-service.js
@@ -43,7 +43,8 @@ experiment('modules/billing/services/volume-matching-service/matching-service', 
 
     test('the matching is aborted early and error codes assigned', async () => {
       expect(result).to.be.an.array().length(1);
-      expect(result[0].calculatedVolume).to.equal(chargeElementGroup.chargeElementContainers[0].chargeElement.volume);
+      expect(result[0].volume).to.equal(chargeElementGroup.chargeElementContainers[0].chargeElement.volume);
+      expect(result[0].calculatedVolume).to.be.undefined();
       expect(result[0].twoPartTariffStatus).to.equal(twoPartTariffStatuses.ERROR_NO_RETURNS_SUBMITTED);
     });
   });
